+++ src/modules/ekf2/EKF/estimator_interface.h
@@ -63,7 +63,7 @@
 
 #include "common.h"
 #include "RingBuffer.h"
-#include "imu_down_sampler.hpp"
+#include "imu_down_sampler/imu_down_sampler.hpp"
 #include "output_predictor/output_predictor.h"
 
 #if defined(CONFIG_EKF2_RANGE_FINDER)
