+++ src/modules/ekf2/EKF/imu_down_sampler/imu_down_sampler.hpp
@@ -41,7 +41,7 @@
 #include <mathlib/mathlib.h>
 #include <matrix/math.hpp>
 
-#include "common.h"
+#include "../common.h"
 
 using namespace estimator;
